feat: add macOS Desktop portable build
Split cmd/cyberagent-desktop by build tag: shared shell code under desktop,
Windows WebView2/Acrylic/registry launcher code under windows, and a new
darwin port. macOS uses bundled WKWebView with no runtime prerequisite probe,
reports startup failures through a bounded strictly escaped osascript dialog,
and launches workspace launchers only through the fixed /usr/bin/open with a
validated .app set. Windows-only seams (Credential Manager, ConPTY terminal,
Safe Web/WFP, controlled host execution) keep failing closed on macOS.

scripts/build-desktop-darwin.sh emits an ad-hoc-signed, un-notarized
build/desktop/Prayu.app with the same consecutive-build SHA-256
reproducibility check, and scripts/check-macos-compat.sh verifies Mach-O,
code signature, metadata, and operator-preview package boundaries. The
renderer reserves the native traffic-light space in the macOS titlebar.
