+++ cmd/cyberagent-desktop/main_darwin_test.go
@@ -1,0 +1,68 @@
+//go:build darwin && desktop
+
+package main
+
+import (
+	"errors"
+	"strings"
+	"testing"
+
+	"cyberagent-workbench/internal/apperror"
+
+	"github.com/wailsapp/wails/v2/pkg/options"
+)
+
+func TestDesktopStartupFailureMessageIsBoundedAndPathFree(t *testing.T) {
+	private := apperror.Wrap(apperror.CodeFailedPrecondition, "database validation failed",
+		errors.New("/Users/private/cyberagent.db"))
+	message := desktopStartupFailureMessage(private)
+	if !strings.Contains(message, string(apperror.CodeFailedPrecondition)) ||
+		strings.Contains(message, "private") || strings.Contains(message, "cyberagent.db") ||
+		len(message) > 256 {
+		t.Fatalf("unsafe startup failure message: %q", message)
+	}
+}
+
+func TestDesktopMacOptionsPinTransparentFramelessBoundary(t *testing.T) {
+	macOptions := desktopMacOptions()
+	if macOptions == nil || macOptions.TitleBar == nil ||
+		!macOptions.TitleBar.HideTitle || !macOptions.TitleBar.FullSizeContent ||
+		!macOptions.WebviewIsTransparent || !macOptions.WindowIsTranslucent {
+		t.Fatalf("macOS frameless boundary is incomplete: %#v", macOptions)
+	}
+	if !macOptions.DisableEscapeExitsFullscreen {
+		t.Fatalf("Escape can exit fullscreen before React modals: %#v", macOptions)
+	}
+	if macOptions.Appearance != "" || macOptions.ContentProtection ||
+		macOptions.OnFileOpen != nil || macOptions.OnUrlOpen != nil {
+		t.Fatalf("macOS options drifted outside the desktop boundary: %#v", macOptions)
+	}
+}
+
+func TestApplyDesktopPlatformOptionsPinsOnlyMacOptions(t *testing.T) {
+	appOptions := &options.App{}
+	applyDesktopPlatformOptions(appOptions)
+	if appOptions.Mac == nil || appOptions.Windows != nil || appOptions.Linux != nil {
+		t.Fatalf("platform options drifted across operating systems: %#v", appOptions)
+	}
+}
+
+func TestCheckDesktopPrerequisitesNeedsNoWebRuntime(t *testing.T) {
+	if err := checkDesktopPrerequisites(); err != nil {
+		t.Fatalf("WKWebView prerequisite check failed: %v", err)
+	}
+}
+
+func TestAppleScriptStringLiteralEscapesStrictly(t *testing.T) {
+	got := appleScriptStringLiteral("line one \"quoted\" \\ path\nline two")
+	want := "\"line one \\\"quoted\\\" \\\\ path\\nline two\""
+	if got != want {
+		t.Fatalf("appleScriptStringLiteral = %q, want %q", got, want)
+	}
+	if strings.Contains(got[1:len(got)-1], "\n") || strings.Contains(got[1:len(got)-1], "\r") {
+		t.Fatalf("unescaped newline escaped into the dialog script: %q", got)
+	}
+	if appleScriptStringLiteral("") != "\"\"" {
+		t.Fatal("empty message did not produce an empty literal")
+	}
+}
