+++ cmd/cyberagent-desktop/main_windows_test.go
@@ -3,152 +3,15 @@
 package main
 
 import (
-	"context"
 	"errors"
-	"io"
-	"net/http"
-	"net/http/httptest"
 	"strings"
 	"testing"
 
 	"cyberagent-workbench/internal/apperror"
-	"cyberagent-workbench/internal/desktop"
 
 	"github.com/wailsapp/wails/v2/pkg/options"
 	"github.com/wailsapp/wails/v2/pkg/options/windows"
 )
-
-type testWindowRestorer struct {
-	unminimised int
-	shown       int
-}
-
-func (r *testWindowRestorer) Unminimise(context.Context) { r.unminimised++ }
-func (r *testWindowRestorer) Show(context.Context)       { r.shown++ }
-
-func newWailsRendererRequest(method, target string, body io.Reader) *http.Request {
-	request := httptest.NewRequest(method, target, body)
-	request.Host = "wails.localhost"
-	request.Header.Set("User-Agent", "PrayuDesktopTest/1.0 wails.io")
-	request.URL.Scheme = ""
-	request.URL.Host = ""
-	request.RequestURI = request.URL.RequestURI()
-	return request
-}
-
-func TestDesktopOptionsDefaultToReadOnlyAndRequireExplicitCapabilities(t *testing.T) {
-	defaults, err := parseDesktopOptions(nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if defaults != (desktopOptions{}) {
-		t.Fatalf("unexpected defaults: %#v", defaults)
-	}
-	for _, current := range []struct {
-		flag string
-		want desktopOptions
-	}{
-		{flag: "--enable-profile-control", want: desktopOptions{profileControl: true}},
-		{flag: "--enable-permission-control", want: desktopOptions{permissionControl: true}},
-		{flag: "--enable-browser-cdp-control", want: desktopOptions{browserCDPControl: true}},
-		{flag: "--enable-run-creation", want: desktopOptions{runCreation: true}},
-		{flag: "--enable-session-messages", want: desktopOptions{sessionMessages: true}},
-		{flag: "--enable-session-steering-control", want: desktopOptions{sessionSteeringControl: true}},
-		{flag: "--enable-run-lifecycle", want: desktopOptions{runLifecycle: true}},
-		{flag: "--enable-run-execution", want: desktopOptions{runExecution: true}},
-		{flag: "--enable-plan-delivery", want: desktopOptions{planDeliveryControl: true}},
-		{flag: "--enable-approvals", want: desktopOptions{approvalControl: true}},
-		{flag: "--enable-command-proposals", want: desktopOptions{commandProposalControl: true}},
-		{flag: "--enable-model-control", want: desktopOptions{modelControl: true}},
-		{flag: "--enable-provider-credentials", want: desktopOptions{providerCredentials: true}},
-		{flag: "--enable-file-edit-review", want: desktopOptions{fileEditReview: true}},
-		{flag: "--enable-file-edit-proposals", want: desktopOptions{fileEditProposals: true}},
-		{flag: "--enable-run-wake", want: desktopOptions{runWakeControl: true}},
-		{flag: "--enable-file-edit-apply", want: desktopOptions{fileEditApply: true}},
-		{flag: "--enable-run-wake-execution", want: desktopOptions{runWakeExecution: true}},
-		{flag: "--enable-wake-worker", want: desktopOptions{runWakeWorker: true}},
-		{flag: "--enable-skill-installation", want: desktopOptions{skillInstallation: true}},
-		{flag: "--enable-evidence-attachments", want: desktopOptions{evidenceAttachment: true}},
-		{flag: "--enable-verification-evidence", want: desktopOptions{verificationEvidence: true}},
-	} {
-		parsed, err := parseDesktopOptions([]string{current.flag})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if parsed != current.want {
-			t.Fatalf("%s was not independently explicit: %#v", current.flag, parsed)
-		}
-	}
-	if _, err := parseDesktopOptions([]string{"unexpected"}); err == nil {
-		t.Fatal("desktop positional argument was accepted")
-	}
-	if _, err := parseDesktopOptions([]string{
-		"--enable-danger-full-access",
-	}); err == nil {
-		t.Fatal("danger-full-access without permission control was accepted")
-	}
-	if _, err := parseDesktopOptions([]string{
-		"--enable-permission-control", "--enable-danger-full-access",
-		"--enable-debug-maximum-access",
-	}); err == nil {
-		t.Fatal("maximum debug access without the user terminal was accepted")
-	}
-	if _, err := parseDesktopOptions([]string{
-		"--enable-full-cdp-debug",
-	}); err == nil {
-		t.Fatal("full CDP debug without its parent gates was accepted")
-	}
-	if _, err := parseDesktopOptions([]string{
-		"--enable-host-command-proposals",
-	}); err == nil {
-		t.Fatal("host command proposals without permission control were accepted")
-	}
-	hostProposals, err := parseDesktopOptions([]string{
-		"--enable-permission-control", "--enable-host-command-proposals",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !hostProposals.permissionControl || !hostProposals.hostCommandProposals {
-		t.Fatalf("host command proposal capability set is incomplete: %+v", hostProposals)
-	}
-	maximum, err := parseDesktopOptions([]string{
-		"--enable-permission-control", "--enable-danger-full-access",
-		"--enable-debug-maximum-access", "--enable-user-terminal",
-		"--enable-browser-cdp-control", "--enable-full-cdp-debug",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !maximum.permissionControl || !maximum.dangerFullAccess ||
-		!maximum.debugMaximumAccess || !maximum.userTerminal ||
-		!maximum.browserCDPControl || !maximum.fullCDPDebug {
-		t.Fatalf("maximum debug capability set is incomplete: %+v", maximum)
-	}
-}
-
-func TestDesktopOperatorPreviewEnablesTheSafeProductBundleOnly(t *testing.T) {
-	preview, err := parseDesktopOptions([]string{"--operator-preview"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !preview.operatorPreview || !preview.profileControl || !preview.permissionControl ||
-		!preview.browserCDPControl || !preview.runCreation || !preview.sessionMessages ||
-		!preview.sessionSteeringControl || !preview.runLifecycle || !preview.runExecution ||
-		!preview.planDeliveryControl || !preview.approvalControl ||
-		!preview.commandProposalControl || !preview.hostCommandProposals || !preview.modelControl ||
-		!preview.providerCredentials || !preview.fileEditReview ||
-		!preview.fileEditProposals || !preview.runWakeControl || !preview.fileEditApply ||
-		!preview.runWakeExecution || !preview.skillInstallation ||
-		!preview.evidenceAttachment || !preview.verificationEvidence ||
-		!preview.embeddedAnalyzer {
-		t.Fatalf("operator preview capability bundle is incomplete: %+v", preview)
-	}
-	if preview.dangerFullAccess || preview.debugMaximumAccess || preview.fullCDPDebug ||
-		preview.runWakeWorker || preview.userTerminal {
-		t.Fatalf("operator preview silently enabled a high-risk capability: %+v", preview)
-	}
-}
 
 func TestDesktopStartupFailureMessageIsBoundedAndPathFree(t *testing.T) {
 	private := apperror.Wrap(apperror.CodeFailedPrecondition, "database validation failed",
@@ -172,6 +35,14 @@ func TestDesktopWindowUsesNativeAcrylicWithoutWeakeningRendererIntegrity(t *test
 		!window.DisablePinchZoom || !window.IsZoomControlEnabled ||
 		window.WindowClassName != "CyberAgentWorkbench" || window.Messages == nil {
 		t.Fatalf("Acrylic changed protected desktop defaults: %#v", window)
+	}
+}
+
+func TestApplyDesktopPlatformOptionsPinsOnlyWindowsOptions(t *testing.T) {
+	appOptions := &options.App{}
+	applyDesktopPlatformOptions(appOptions)
+	if appOptions.Windows == nil || appOptions.Mac != nil || appOptions.Linux != nil {
+		t.Fatalf("platform options drifted across operating systems: %#v", appOptions)
 	}
 }
 
@@ -226,259 +97,5 @@ func TestWebView2PrerequisiteFailsClosedWithoutStartingAnInstaller(t *testing.T)
 		strings.Contains(all, "silently") || strings.Contains(all, "press ok") ||
 		messages.DownloadPage != "" || messages.PressOKToInstall != "" {
 		t.Fatalf("WebView2 messages can trigger or direct an implicit installer: %q", all)
-	}
-}
-
-func TestSecondInstanceHandlerIgnoresArgumentsAndRecoversAfterStartup(t *testing.T) {
-	restorer := &testWindowRestorer{}
-	lifecycle := desktop.NewLifecycle(restorer)
-	handler := secondInstanceHandler(lifecycle)
-	handler(options.SecondInstanceData{
-		Args: []string{"--secret", `C:\PRIVATE\workspace`}, WorkingDirectory: `C:\PRIVATE`,
-	})
-	if restorer.unminimised != 0 || restorer.shown != 0 {
-		t.Fatal("second instance restored the window before startup")
-	}
-	lifecycle.Start(context.Background())
-	if restorer.unminimised != 1 || restorer.shown != 1 {
-		t.Fatalf("second instance recovery count = %d/%d", restorer.unminimised, restorer.shown)
-	}
-	lifecycle.Stop()
-}
-
-func TestInProcessAPIHandlerPinsLoopbackBoundaryWithoutMutatingRequest(t *testing.T) {
-	var receivedHost string
-	var receivedRemote string
-	var receivedPath string
-	var receivedScheme string
-	var receivedURLHost string
-	next := http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		receivedHost = request.Host
-		receivedRemote = request.RemoteAddr
-		receivedPath = request.URL.Path
-		receivedScheme = request.URL.Scheme
-		receivedURLHost = request.URL.Host
-		writer.WriteHeader(http.StatusNoContent)
-	})
-	handler := inProcessAPIHandler{next: next}
-	request := newWailsRendererRequest(http.MethodGet, "http://wails.localhost/api/v1/health", nil)
-	request.RemoteAddr = "203.0.113.10:443"
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-
-	if response.Code != http.StatusNoContent || receivedHost != "127.0.0.1" ||
-		receivedRemote != "127.0.0.1:0" || receivedPath != "/api/v1/health" ||
-		receivedScheme != "" || receivedURLHost != "" {
-		t.Fatalf("unexpected in-process projection: status=%d host=%q remote=%q path=%q",
-			response.Code, receivedHost, receivedRemote, receivedPath)
-	}
-	if request.Host != "wails.localhost" || request.RemoteAddr != "203.0.113.10:443" ||
-		request.URL.Scheme != "" || request.URL.Host != "" {
-		t.Fatalf("original request was mutated: host=%q remote=%q", request.Host, request.RemoteAddr)
-	}
-}
-
-func TestInProcessAPIHandlerAcceptsWailsWebViewRequestShape(t *testing.T) {
-	called := false
-	handler := inProcessAPIHandler{next: http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		called = true
-		writer.WriteHeader(http.StatusNoContent)
-	})}
-	request := newWailsRendererRequest(http.MethodGet,
-		"http://wails.localhost/api/v1/desktop/bootstrap", nil)
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusNoContent || !called {
-		t.Fatalf("real Wails request shape was rejected: status=%d called=%t", response.Code, called)
-	}
-}
-
-func TestInProcessAPIHandlerRejectsNonRendererOrigins(t *testing.T) {
-	for _, target := range []string{
-		"https://wails.localhost/api/v1/health",
-		"http://wails.localhost:80/api/v1/health",
-		"http://user@wails.localhost/api/v1/health",
-		"http://untrusted.example/api/v1/health",
-	} {
-		t.Run(target, func(t *testing.T) {
-			called := false
-			handler := inProcessAPIHandler{next: http.HandlerFunc(func(http.ResponseWriter, *http.Request) {
-				called = true
-			})}
-			response := httptest.NewRecorder()
-			request := httptest.NewRequest(http.MethodGet, target, nil)
-			request.Header.Set("User-Agent", "PrayuDesktopTest/1.0 wails.io")
-			handler.ServeHTTP(response, request)
-			if response.Code != http.StatusForbidden || called {
-				t.Fatalf("origin %q reached API: status=%d called=%t", target, response.Code, called)
-			}
-		})
-	}
-}
-
-func TestInProcessAPIHandlerRejectsNonCanonicalRendererURL(t *testing.T) {
-	tests := []struct {
-		name   string
-		mutate func(*http.Request)
-	}{
-		{name: "fragment", mutate: func(request *http.Request) { request.URL.Fragment = "fragment" }},
-		{name: "opaque", mutate: func(request *http.Request) {
-			request.URL.Opaque = "//wails.localhost/api/v1/health"
-		}},
-	}
-	for _, current := range tests {
-		t.Run(current.name, func(t *testing.T) {
-			called := false
-			handler := inProcessAPIHandler{next: http.HandlerFunc(func(http.ResponseWriter, *http.Request) {
-				called = true
-			})}
-			request := newWailsRendererRequest(http.MethodGet,
-				"http://wails.localhost/api/v1/health", nil)
-			current.mutate(request)
-			response := httptest.NewRecorder()
-			handler.ServeHTTP(response, request)
-			if response.Code != http.StatusForbidden || called {
-				t.Fatalf("non-canonical renderer URL reached API: status=%d called=%t",
-					response.Code, called)
-			}
-		})
-	}
-}
-
-func TestInProcessAPIHandlerRejectsOriginlessRequestsWithoutWailsAuthority(t *testing.T) {
-	tests := []struct {
-		name   string
-		mutate func(*http.Request)
-	}{
-		{name: "missing user agent", mutate: func(request *http.Request) {
-			request.Header.Del("User-Agent")
-		}},
-		{name: "substring user agent", mutate: func(request *http.Request) {
-			request.Header.Set("User-Agent", "not-wails.io-client")
-		}},
-		{name: "missing host", mutate: func(request *http.Request) { request.Host = "" }},
-		{name: "wrong host", mutate: func(request *http.Request) { request.Host = "untrusted.example" }},
-		{name: "host port", mutate: func(request *http.Request) { request.Host = "wails.localhost:80" }},
-		{name: "host whitespace", mutate: func(request *http.Request) { request.Host = " wails.localhost" }},
-		{name: "partial authority", mutate: func(request *http.Request) { request.URL.Scheme = "http" }},
-	}
-	for _, current := range tests {
-		t.Run(current.name, func(t *testing.T) {
-			called := false
-			handler := inProcessAPIHandler{next: http.HandlerFunc(func(http.ResponseWriter, *http.Request) {
-				called = true
-			})}
-			request := newWailsRendererRequest(http.MethodGet,
-				"http://wails.localhost/api/v1/health", nil)
-			current.mutate(request)
-			response := httptest.NewRecorder()
-			handler.ServeHTTP(response, request)
-			if response.Code != http.StatusForbidden || called {
-				t.Fatalf("untrusted originless request reached API: status=%d called=%t",
-					response.Code, called)
-			}
-		})
-	}
-}
-
-func TestInProcessAPIHandlerFailsClosedWithoutTarget(t *testing.T) {
-	request := httptest.NewRequest(http.MethodGet, "http://wails.localhost/", nil)
-	response := httptest.NewRecorder()
-	inProcessAPIHandler{}.ServeHTTP(response, request)
-	if response.Code != http.StatusInternalServerError {
-		t.Fatalf("status = %d, want %d", response.Code, http.StatusInternalServerError)
-	}
-}
-
-func TestInProcessAPIHandlerFailsClosedWithoutURL(t *testing.T) {
-	request := httptest.NewRequest(http.MethodGet, "http://wails.localhost/", nil)
-	request.URL = nil
-	response := httptest.NewRecorder()
-	inProcessAPIHandler{next: http.HandlerFunc(func(http.ResponseWriter, *http.Request) {
-		t.Fatal("nil URL request reached the API")
-	})}.ServeHTTP(response, request)
-	if response.Code != http.StatusInternalServerError {
-		t.Fatalf("status = %d, want %d", response.Code, http.StatusInternalServerError)
-	}
-}
-
-func TestInProcessAPIHandlerRebuildsMissingRequestURIFromURL(t *testing.T) {
-	var requestURI string
-	handler := inProcessAPIHandler{next: http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		requestURI = request.RequestURI
-		writer.WriteHeader(http.StatusNoContent)
-	})}
-	request := newWailsRendererRequest(http.MethodGet,
-		"http://wails.localhost/api/v1/health?probe=one", nil)
-	request.RequestURI = ""
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusNoContent || requestURI != "/api/v1/health?probe=one" {
-		t.Fatalf("status=%d request_uri=%q", response.Code, requestURI)
-	}
-	if request.RequestURI != "" {
-		t.Fatal("source request URI was mutated")
-	}
-}
-
-func TestInProcessAPIHandlerCanonicalizesMismatchedRequestURI(t *testing.T) {
-	var requestURI string
-	handler := inProcessAPIHandler{next: http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		requestURI = request.RequestURI
-		writer.WriteHeader(http.StatusNoContent)
-	})}
-	request := newWailsRendererRequest(http.MethodGet,
-		"http://wails.localhost/api/v1/health?probe=one", nil)
-	request.RequestURI = "http://untrusted.example/private?secret=true"
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusNoContent || requestURI != "/api/v1/health?probe=one" {
-		t.Fatalf("status=%d request_uri=%q", response.Code, requestURI)
-	}
-	if request.RequestURI != "http://untrusted.example/private?secret=true" {
-		t.Fatal("source request URI was mutated")
-	}
-}
-
-func TestInProcessAPIHandlerCanonicalizesOnlyTheWailsEmptyRoot(t *testing.T) {
-	var path string
-	var requestURI string
-	var contentLength int64
-	handler := inProcessAPIHandler{next: http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		path = request.URL.Path
-		requestURI = request.RequestURI
-		contentLength = request.ContentLength
-		writer.WriteHeader(http.StatusNoContent)
-	})}
-	request := newWailsRendererRequest(http.MethodGet, "http://wails.localhost/", nil)
-	request.URL.Path = ""
-	request.RequestURI = ""
-	request.ContentLength = -1
-	request.Body = http.NoBody
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusNoContent || path != "/" || requestURI != "/" || contentLength != 0 {
-		t.Fatalf("status=%d path=%q request_uri=%q content_length=%d",
-			response.Code, path, requestURI, contentLength)
-	}
-	if request.URL.Path != "" || request.RequestURI != "" || request.ContentLength != -1 {
-		t.Fatal("Wails source request was mutated")
-	}
-}
-
-func TestInProcessAPIHandlerDoesNotEraseUnknownRequestBodies(t *testing.T) {
-	var contentLength int64
-	handler := inProcessAPIHandler{next: http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		contentLength = request.ContentLength
-		writer.WriteHeader(http.StatusNoContent)
-	})}
-	request := newWailsRendererRequest(http.MethodGet,
-		"http://wails.localhost/api/v1/health", nil)
-	request.ContentLength = -1
-	request.Body = io.NopCloser(strings.NewReader("unexpected"))
-	response := httptest.NewRecorder()
-	handler.ServeHTTP(response, request)
-	if response.Code != http.StatusNoContent || contentLength != -1 {
-		t.Fatalf("status=%d content_length=%d", response.Code, contentLength)
 	}
 }
